requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
PyJWT==2.10.1
cryptography==45.0.6
argon2-cffi==25.1.0
redis==6.4.0
pytest==8.4.1
pytest-asyncio==1.1.0
